server actions are now secure

+++ lib/capsuleRelatedServerActions.ts
@@ -11,6 +11,10 @@ import {
     CapsuleWithUsers,
 } from '@/lib/types';
 import { getPartnerFromUser } from '@/lib/db_utils';
+import { getServerSession } from 'next-auth/next';
+import { authOptions } from '@/auth';
+import { getUserWithPartnershipByEmail } from '@/lib/partnerRequestServerActions';
+
 // import { getUserWithPartnershipByEmail } from '@/lib/capsuleRelatedServerActions';
 
 export async function createCapsule(
@@ -19,6 +23,15 @@ export async function createCapsule(
     color: string | null,
     message: string,
 ) {
+    const session = await getServerSession(authOptions);
+    if (!session?.user) {
+        throw new Error('No user found in session');
+    }
+    if (session.user.email !== user.email) {
+        throw new Error(
+            `User in session (${session.user.email}) does not match user in request (${user.email})`,
+        );
+    }
     const data = {
         message: message,
         color: color || randColor(),
@@ -45,7 +58,16 @@ export async function createCapsule(
     revalidatePath(path);
 }
 
-export async function deleteCapsule(path: string, capsule: Capsule) {
+export async function deleteCapsule(path: string, capsule: CapsuleWithUsers) {
+    const session = await getServerSession(authOptions);
+    if (!session?.user) {
+        throw new Error('No user found in session');
+    }
+    if (session.user.email !== capsule.author.email) {
+        throw new Error(
+            `User in session (${session.user.email}) does not match user in request (${capsule.author.email})`,
+        );
+    }
     console.log('DELETING CAPSULE: ', capsule);
     const response = await prisma.capsule.delete({
         where: {
@@ -58,10 +80,20 @@ export async function deleteCapsule(path: string, capsule: Capsule) {
 
 export async function updateCapsuleScalars(
     path: string,
-    capsule: Capsule,
+    capsule: CapsuleWithUsers,
     color: string | null,
     message: string,
 ) {
+    const session = await getServerSession(authOptions);
+    if (!session?.user) {
+        throw new Error('No user found in session');
+    }
+    if (session.user.email !== capsule.author.email) {
+        throw new Error(
+            `User in session (${session.user.email}) does not match user in request (${capsule.author.email})`,
+        );
+    }
+
     const data: any = {
         message: message,
         ...(!!color && { color: color }),
@@ -78,11 +110,30 @@ export async function updateCapsuleScalars(
 
 export async function updateCapsuleOpen(
     path: string,
-    capsule: Capsule,
+    capsule: CapsuleWithUsers,
     user: UserWithPartnershipAndAuthoredCapsules,
     open: boolean = true,
     revalidate: boolean = true,
 ) {
+    const session = await getServerSession(authOptions);
+    if (!session?.user) {
+        throw new Error('No user found in session');
+    }
+    if (user.email !== session.user.email) {
+        throw new Error(
+            `User in session (${session.user.email}) does not match user in request (${user.email})`,
+        );
+    }
+    if (
+        !(capsule?.partnership?.partners || []).some(
+            (partner) => partner.email === user.email,
+        )
+    ) {
+        throw new Error(
+            `User ${user.email} is not a partner of capsule ${capsule.id}`,
+        );
+    }
+
     const response = await prisma.capsule.update({
         where: {
             id: capsule.id,
@@ -107,9 +158,32 @@ export async function updateCapsuleOpen(
 }
 
 export async function sealCapsule(path: string, capsule: CapsuleWithUsers) {
+    const session = await getServerSession(authOptions);
+    if (!session?.user?.email) {
+        throw new Error('No user found in session');
+    }
     if (!capsule.author.partnershipId) {
         throw new Error('Cannot seal a capsule without a partnership');
     }
+
+    const user = await getUserWithPartnershipByEmail(session.user.email);
+    if (capsule.author.email !== session.user.email) {
+        if (!capsule.partnershipId) {
+            throw new Error(
+                `User ${session?.user?.email} is not a author or in partnership of capsule ${capsule.id}`,
+            );
+        }
+        if (
+            !(capsule?.partnership?.partners || []).some(
+                (partner) => partner.email === session?.user?.email,
+            )
+        ) {
+            throw new Error(
+                `User ${session?.user?.email} is not a partner of capsule ${capsule.id}`,
+            );
+        }
+    }
+
     const response = await prisma.capsule.update({
         where: {
             id: capsule.id,
